requirements.txt
contourpy==1.1.1
cycler==0.12.1
distlib==0.3.9
filelock==3.16.1
importlib-resources==6.4.5
kiwisolver==1.4.7
matplotlib==3.7.5
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pillow==10.4.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
tzdata==2024.2
zipp==3.20.2
